DevOps/microservice/oauth2/pyproject.toml
[tool.poetry]
name = "oauth2"
version = "0.1.0"
description = ""
authors = ["Sarim95khan <sarim95khan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
